convert whitebox test to blackbox now that there is an accessor, rename file+testfile
