--- tools/migration/internal/get_repo_paths_test.go
@@ -1,33 +1,0 @@
-package internal
-
-import (
-	"regexp"
-	"testing"
-
-	ast "github.com/stretchr/testify/assert"
-	req "github.com/stretchr/testify/require"
-
-	tf "github.com/filecoin-project/go-filecoin/testhelpers/testflags"
-)
-
-func TestGetNewRepoPath(t *testing.T) {
-	tf.UnitTest(t)
-	assert := ast.New(t)
-	require := req.New(t)
-
-	t.Run("Uses the new repo opt as a prefix if provided", func(t *testing.T) {
-		dirname := "/tmp/mynew_repodir"
-		newpath := getNewRepoPath("/tmp/myold_repodir", dirname, "1", "2")
-		rgx, err := regexp.Compile("/tmp/mynew_repodir_1_2_[0-9]{8}-[0-9]{6}$")
-		require.NoError(err)
-		assert.Regexp(rgx, newpath)
-	})
-
-	t.Run("Adds a timestamp to the new repo dir", func(t *testing.T) {
-		dirname := "/tmp/myfilecoindir"
-		newpath := getNewRepoPath(dirname, "", "1", "2")
-		rgx, err := regexp.Compile("/tmp/myfilecoindir_1_2_[0-9]{8}-[0-9]{6}$")
-		require.NoError(err)
-		assert.Regexp(rgx, newpath)
-	})
-}
